rename: change directory name (contents -> data)
