+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import { Footer } from '@/components/footer';
 import { Header } from '@/components/header';
 import { ThemeProvider } from '@/components/theme';
-import meta from '@/contents/metadata.json';
+import meta from '@/data/metadata.json';
 import '@/styles/globals.css';
 import { jetBrainsMono, pretendard } from '@/utils/font';
 import type { Metadata } from 'next';
