+++ src/data/posts/auto-deployment.mdx
